requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
plotly>=5.0.0
pytesseract>=0.3.10
Pillow>=9.0.0
pathlib